refactor(Frames): 일급 콜렉션 만들기
